% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_large_strips.R
\name{add_large_strips}
\alias{add_large_strips}
\title{Add large strips to facet plot.}
\usage{
add_large_strips(
  p1,
  top.strip,
  right.strip,
  strip.fill = "gray85",
  strip.col = NA
)
}
\arguments{
\item{top.strip}{Char for label of top strip.}

\item{right.strip}{Char for label of right strip.}

\item{strip.fill}{Colour value for fill of strips. (Default = "gray85")}

\item{strip.col}{Colour value for colour of ourtline for strip. Defatults to "NA".}
}
\description{
Adds 'overall' facets for the top and right of a facet plot.
}
\examples{
p2 <- ggplot(CO2, aes(conc, uptake, colour = Plant)) +
geom_point() +
facet_grid(cols=vars(Type), rows = vars(Treatment))

add_large_strips(p2, top.strip = "Plant Type", right.strip = "Treatment")
}
\keyword{facet}
\keyword{ggplot}
\keyword{strip}
