% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_light.R
\name{simple_light}
\alias{simple_light}
\title{Simple light theme.}
\usage{
simple_light(AR = 1/2, legend = TRUE, linesize = 1)
}
\arguments{
\item{AR}{Aspect ratio of plot as dbl.}

\item{legend}{Whether or not to include legend, TRUE by default.}

\item{linesize}{Thickness of plot border, defaults to 1.}
}
\description{
This function strips the specified axis from the plot by supplying element_blank() for all of the relevant theme components.
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) + geom_point() + strip_axis(axis = "x")
}
\keyword{ggplot}
\keyword{strip,}
