% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_axis.R
\name{strip_axis}
\alias{strip_axis}
\title{Strip an axis.}
\usage{
strip_axis(axis = "y")
}
\arguments{
\item{axis}{Which axis to strip, either x or y.}
}
\description{
This function strips the specified axis from the plot by supplying element_blank() for all of the relevant theme components.
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) + geom_point() + strip_axis(axis = "x")
}
\keyword{ggplot}
\keyword{strip,}
