% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dark.R
\name{simple_dark}
\alias{simple_dark}
\title{A Dark Theme for ggplot2}
\usage{
simple_dark(linesize = 2)
}
\arguments{
\item{love}{Aspect ratio of plot, defaults to 1/2.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) + geom_point() + simple_dark()
}
\keyword{darj}
\keyword{darkmode,}
\keyword{ggplot,}
\keyword{theme,}
